// src/app/models/enums.ts
export enum Gender {
  Male = 'male',
  Female = 'female'
}

export enum Citizenship {
  USA = 'United States of America',
  Ireland = 'Ireland',
  Wakanda = 'Wakanda',
  UK = 'United Kingdom',
  Canada = 'Canada',
  Wallachia = 'Principality of Wallachia',
  ShiAr = "Shi'ar",
  Russia = 'Russia',
  Egypt = 'Egypt',
  Asgard = 'Asgard'
}

export enum Skill {
  SuperhumanStrength = 'superhuman strength',
  SonicScream = 'sonic scream',
  Precognition = 'precognition',
  HealingFactor = 'healing factor',
  SuperhumanAgility = 'superhuman agility',
  Reflexes = 'reflexes',
  Telepathy = 'telepathy',
  EnergyBlasts = 'energy blasts',
  Photokinesis = 'photokinesis',
  Shapeshifting = 'shapeshifting',
  WeatherManipulation = 'weather manipulation',
  Levitation = 'levitation',
  Retrocognition = 'retrocognition',
  TeleportationInFiction = 'teleportation in fiction',
}

export enum Occupation {
  Psychologist = 'psychologist',
  Student = 'student',
  Criminal = 'criminal',
  Scientist = 'scientist',
  Writer = 'writer',
  Superhero = 'superhero',
  Soldier = 'soldier',
  FictionalSoldier = 'fictional soldier',
  Sovereign = 'sovereign',
  Businessperson = 'businessperson',
  CircusPerformer = 'circus performer',
  Leader = 'leader',
  Teacher = 'teacher',
  Reporter = 'reporter',
  Actor = 'actor',
  Terrorist = 'terrorist',
  Gangster = 'gangster',
  Mercenary = 'mercenary',
  DomesticWorker = 'domestic worker',
  Astronaut = 'astronaut',
  Physician = 'physician',
  Sailor = 'sailor'
}

export enum MemberOf {
  HorsemenOfApocalypse = 'Horsemen of Apocalypse',
  TheSpiderSociety = 'The Spider Society',
  Interpol = 'Interpol',
  NewWarriors = 'New Warriors',
  Defenders = 'Defenders',
  TheAvengers = 'The Avengers',
  BrotherhoodOfMutants = 'Brotherhood of Mutants',
  UnitedStatesArmy = 'United States Army',
  Illuminati = 'Illuminati',
  HellfireClub = 'Hellfire Club',
  XMen = 'X-Men',
  OrderOfTheDragon = 'Order of the Dragon',
  Thunderbolts = 'Thunderbolts',
  ImperialGuard = 'Imperial Guard',
  Alchemax = 'Alchemax',
  Avengers = 'Avengers',
  Purifiers = 'Purifiers',
  DarkAvengers = 'Dark Avengers',
  KGB = 'KGB',
  ClanAkkaba = 'Clan Akkaba',
  Oscorp = 'Oscorp',
  SHIELD = 'S.H.I.E.L.D.',
  XForce = 'X-Force',
  CanadianArmy = 'Canadian Army',
  AlphaFlight = 'Alpha Flight',
  Maggia = 'Maggia',
  UniversityOfCambridge = 'University of Cambridge',
  FantasticFour = 'Fantastic Four'
}

export enum Creator {
  WaltSimonson = 'Walt Simonson',
  FionaAvery = 'Fiona Avery',
  RoyThomas = 'Roy Thomas',
  GerryConway = 'Gerry Conway',
  StanLee = 'Stan Lee',
  JackKirby = 'Jack Kirby',
  ChrisClaremont = 'Chris Claremont',
  GrantMorrison = 'Grant Morrison',
  FabianNicieza = 'Fabian Nicieza',
  GaryFriedrich = 'Gary Friedrich',
  JohnRomitaJr = 'John Romita Jr.',
  BrentAnderson = 'Brent Anderson',
  StevenButler = 'Steven Butler',
  JohnByrne = 'John Byrne',
  RonFrenz = 'Ron Frenz',
  BrianMichaelBendis = 'Brian Michael Bendis',
  MarkMillar = 'Mark Millar',
  MarkBagley = 'Mark Bagley',
  LenWein = 'Len Wein',
  JimLee = 'Jim Lee',
  ScottLobdell = 'Scott Lobdell',
  SteveDitko = 'Steve Ditko'
}